% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_helpers.R
\name{add_analyte_scores}
\alias{add_analyte_scores}
\title{Add analyte scores to MOModules Object.}
\usage{
add_analyte_scores(scores_df = NULL, name = NULL, map = TRUE)
}
\arguments{
\item{scores_df}{Data.frame of scores including the following columns: AnalyteID (name of analyte), AnalyteIDtype (type of analyte for processing), Assay (one of `Proteomics`, `Metabolomics` or `Transcriptomics`), Score (a numeric), Direction (either `Pos` for Positive or `Neg` for Negative), and Significant (boolean, `TRUE` or `FALSE`)}

\item{name}{What to store the `scores_df` under (see `ModObj$scores$...`)? Defaults to `table1` for first scores, `table2`, and so on.}

\item{map}{Try to map the data to the multi-omics modules? Defaults to `TRUE`}
}
\description{
Add analyte scores to MOModules Object.
}
\examples{
ModObj <- make.multiomics.modules(...)

ModObj$get_module_scores(...)

}
