% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_helpers.R
\name{convert_id}
\alias{convert_id}
\title{Convert an analyte id to be used within the IMPACC scores:}
\usage{
convert_id(analyte.id, analyte.id.type, assay)
}
\arguments{
\item{analyte.id}{ID of analyte to convert.}

\item{analyte.id.type}{Type of ID (e.g. ENSEMBL, HMBD, KEGG, GENE_SYMBOL, ...)}

\item{assay}{Which assay? (Proteomics, Transcriptomics, or Metabolomics)}
}
\description{
Convert an analyte id to be used within the IMPACC scores:
}
\examples{
ModObj$convert_id("IL6", "GENE_SYMBOL", "Proteomics")
 
}
