% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_helpers.R
\name{do_enrichment}
\alias{do_enrichment}
\title{Perform enrichment analysis using the determined method on the chosen scores data.frame.}
\usage{
do_enrichment(name = "table1", method = "match", return = TRUE)
}
\arguments{
\item{name}{What version of `scores_df` to use (see `ModObj$scores$...`)? Added via `$add_analyte_scores(...)`. Defaults to `"table1"`.}

\item{method}{What method of enrichment to use? Defaults to `"match"`}

\item{return}{Return the table stored under `$reports$...`? Defaults to `TRUE`}
}
\description{
Perform enrichment analysis using the determined method on the chosen scores data.frame.
}
\examples{
ModObj$do_enrichment()

}
