% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_helpers.R
\name{get_module_scores}
\alias{get_module_scores}
\title{Get multi-omics module scores for a list of analytes.}
\usage{
get_module_scores(name = NULL, method = "gsea", correction = "BH", seed = 42)
}
\arguments{
\item{name}{What version of `scores_df` to use (see `ModObj$scores$...`)? Added via `$add_analyte_scores(...)`. Defaults to `"table1"`.}

\item{method}{What method to derive scores? Defaults to `"gsea"`.}

\item{correction}{How to correct p.values (if method == `"gsea"` or `"fischer"`)? Defaults to `"BH"`. Use `"none"` to prevent correction.}

\item{seed}{Seed to set (for `method = "gsea"`). Defaults to `42`.}
}
\description{
Get multi-omics module scores for a list of analytes.
}
\examples{
get_module_scores(...)

}
