% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ModulesObject.R
\name{make.multiomics.modules}
\alias{make.multiomics.modules}
\title{Make a Multi-omics Modules object. Will return an R6 class MOModules used for the "MOModules" package.}
\usage{
make.multiomics.modules(quiet = FALSE)
}
\arguments{
\item{quiet}{Mute print statements? Defaults to `FALSE`.}
}
\description{
Make a Multi-omics Modules object. Will return an R6 class MOModules used for the "MOModules" package.
}
