% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_helpers.R
\name{plot_fingerprint}
\alias{plot_fingerprint}
\title{Plot the representative analytes of each assay (Proteomics, Metabolomics, Transcriptomics) vs scores}
\usage{
plot_fingerprint(
  name = NULL,
  module = 1,
  use = "dir",
  invert = FALSE,
  only.show.sig = TRUE,
  all.assays = FALSE,
  show.analyte.names = FALSE
)
}
\arguments{
\item{name}{What version of `scores_df` to use (see `ModObj$scores$...`)? Added via `$add_analyte_scores(...)`. Defaults to all (`NULL`).}

\item{module}{Which module to plot (as integer)? Defaults to `1`.}

\item{use}{Which variable to plot for Data? Can be `"score"` (score provided) or `"dir"` (direction)}

\item{invert}{Should the module directions be inverted? Defaults to `FALSE`}

\item{only.show.sig}{Only show significant analytes. Defaults to `TRUE`}

\item{all.assays}{Show assays even if no analytes are present? Defaults to `FALSE`}

\item{show.analyte.names}{Include x-axis names (analytes)? Defaults to `FALSE`}
}
\description{
Plot the representative analytes of each assay (Proteomics, Metabolomics, Transcriptomics) vs scores
}
\examples{
ModObj$plot_mapping()

ModObj$plot_mapping(name = "my_scores")

}
