% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_helpers.R
\name{plot_mapping}
\alias{plot_mapping}
\title{Plot the overlap per analyte of each assay (Proteomics, Metabolomics, Transcriptomics)}
\usage{
plot_mapping(name = "table1")
}
\arguments{
\item{name}{What version of `scores_df` to use (see `ModObj$scores$...`)? Added via `$add_analyte_scores(...)`. Defaults to `"table1"`.}
}
\description{
Plot the overlap per analyte of each assay (Proteomics, Metabolomics, Transcriptomics)
}
\examples{
ModObj$plot_mapping()

ModObj$plot_mapping(name = "my_scores")

}
