% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_helpers.R
\name{plot_module_scores}
\alias{plot_module_scores}
\title{Plot the representative analytes of each assay (Proteomics, Metabolomics, Transcriptomics) vs scores}
\usage{
plot_module_scores(name = NULL, assay = "Total")
}
\arguments{
\item{name}{What version of `scores_df` to use (see `ModObj$scores$...`)? Added via `$add_analyte_scores(...)`. Defaults to all (`NULL`).}

\item{assay}{What assay to show? Can be `"total"` (default) or `"proteomics"`, `"metabolomics"`, or `"transcriptomics"`.}
}
\description{
Plot the representative analytes of each assay (Proteomics, Metabolomics, Transcriptomics) vs scores
}
\examples{
# run first:
ModObj$get_module_scores(...)
ModObj$plot_module_scores()

}
