% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_helpers.R
\name{plot_report}
\alias{plot_report}
\title{Plot the representative analytes of each assay (Proteomics, Metabolomics, Transcriptomics) vs scores}
\usage{
plot_report(
  name = NULL,
  module = NULL,
  assay = NULL,
  type = "gsea",
  cap = 1e-20,
  map.cutoff = 0.15,
  show.adjust = TRUE,
  show.all = FALSE
)
}
\arguments{
\item{name}{What version of `scores_df` to use (see `ModObj$scores$...`)? Added via `$add_analyte_scores(...)`. Defaults to all (`NULL`).}

\item{module}{What modules to plot? Defaults to all (`NULL`).}

\item{assay}{What assays to plot? Defaults to all (`NULL`).}

\item{type}{Type of report (generated from `get_module_scores`), defaults to `"gsea"`.}

\item{cap}{Should p-value magnitudes be capped? Useful when a few very significant p-values drive the visualization. Defaults to `1e-5`}

\item{map.cutoff}{Required proportion of analytes to be mapped? Defaults to `0.15` (15%). Use `0` to ignore cutoff.}

\item{show.adjust}{Show adjusted p.values? Defaults to `TRUE`.}

\item{show.all}{Show all p.values? Defaults to `FALSE`.}
}
\description{
Plot the representative analytes of each assay (Proteomics, Metabolomics, Transcriptomics) vs scores
}
\examples{
# run first:
ModObj$get_module_scores(...)
ModObj$plot_module_scores()

}
