% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_helpers.R
\name{get_module_scores}
\alias{get_module_scores}
\title{Get multi-omics module scores for a list of analytes.}
\usage{
get_module_scores(
  name = NULL,
  method = "fisher",
  p.adjust = "BH",
  min.analytes = 0
)
}
\arguments{
\item{name}{What version of `scores_df` to use (see `ModObj$scores$...`)? Added via `$add_analyte_scores(...)`. Defaults to `"table1"`.}

\item{method}{What method to derive scores? Defaults to `"fisher"`.}

\item{p.adjust}{Method to adjust p.values? Defaults to `"BH"`. Can also be `"none"`.}

\item{min.analytes}{Minimum number of analytes needed to calculate a score (per assay). Defaults to `0`.}
}
\description{
Get multi-omics module scores for a list of analytes.
}
\examples{
get_module_scores(...)

}
