% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_helpers.R
\name{plot_fingerprint}
\alias{plot_fingerprint}
\title{Plot the representative analytes of each assay (Proteomics, Metabolomics, Transcriptomics) vs scores}
\usage{
plot_fingerprint(
  name = NULL,
  assay = NULL,
  module = 1,
  use = "p.value",
  cap = NULL,
  invert = FALSE,
  only.show.sig = TRUE,
  show.analyte.names = TRUE
)
}
\arguments{
\item{name}{What version of `scores_df` to use (see `ModObj$scores$...`)? Added via `$add_analyte_scores(...)`. Defaults to all (`NULL`).}

\item{assay}{What assays to plot? Defaults to all (`NULL`).}

\item{module}{Which module to plot (as integer)? Defaults to `1`.}

\item{use}{Which variable to plot for Data? Can be `"p.value"` (p.value, default), `"score"` (score provided) or `"dir"` (direction)}

\item{cap}{Provide a cap for the plot? Defaults to `NULL` (no cap). Alternatively, provide a numeric.}

\item{invert}{Should the module directions be inverted? Defaults to `FALSE`}

\item{only.show.sig}{Only show significant analytes. Defaults to `TRUE`}

\item{show.analyte.names}{Include x-axis names (analytes)? Defaults to `TRUE`}
}
\description{
Plot the representative analytes of each assay (Proteomics, Metabolomics, Transcriptomics) vs scores
}
\examples{
ModObj$plot_mapping()

ModObj$plot_mapping(name = "my_scores")

}
